Browse table returned by cjpg_table() or cjsg_table()

Browse table returned by cjpg_table() or cjsg_table()

browse_table(table, patterns)

Arguments

table
Table returned by cjpg_table() or cjsg_table() (only valid for "classes" or "subjects" types)

patterns
A list containing (at most) 6 character vectors of one or more regular expressions (applied from left to right on root to leaves); e.g. list(c("ADM", "CRIMINAL"), "", "", "", "", "")

Details

Regex of the same level will be ORed and of different levels will be ANDed, e.g., list(c("ADM", "CRIMINAL"), "", "", "", "", "") becomes list("(?:ADM|CRIMINAL)", "", "", "", "", "") and each element will be applied with dplyr::filter() to a level of the tree

See also

cjpg_table(), cjsg_table()