Download second degree lawsuits filed in Brazilian Justice Courts

This function downloads lawsuits as PDFs. Given a lawsuit ID, and the path to a directory it will collect the lawsuit, and save it to the provided directory.

download_2deg_lawsuit(id, path = ".")

Arguments

id
A character vector of one or more lawsuit IDs (only works with TJSP for now)

path
Path to the directory where the lawsuit should be saved

Value

A character vector with the path to the downloaded lawsuit

About lawsuits

The lawsuits contemplated by this funtion have to be filed in a Brazilian Tribunal de Justica (Justice Court). download_2deg_lawsuit() finds the lawsuit in its state's online Sistema de Automacao de Justica (Justice Automation System), solves the captcha withholding the information, and collects the PDF.

Implemented TJs

Unfortunatelly download_2deg_lawsuit() doesn't yet work with all 27 TJs in Brazil. Here are the ones already implemented:

TJSP (Sao Paulo)