Download results of a query on second degree lawsuits filed in Brazilian Justice Courts

Downloads an HTML with the results obtained from querying a dataset of all second degree lawsuits and then one HTML for each page of results (at most max_page pages). query should be the string to look for in the lawsuits and clases, courts, etc. should be the filtering parameters (make sure to use cjsg_table() to get lists of all valid codes for these arguments).

download_cjsg(query, path = ".", classes = "", subjects = "",
  courts = "", trial_start = "", trial_end = "",
  registration_start = "", registration_end = "", min_page = 1,
  max_page = 1, cores = 1, tj = "tjsp")

Arguments

query
Character vector with search query

path
Path to directory where to save HTMLs

classes
Character vector with class IDs (see cjsg_table())

subjects
Character vector with subject IDs (see cjsg_table())

courts
Character vector with court IDs (see cjsg_table())

min_page
First page of results to download

max_page
Last page of results to download

cores
The number of cores to be used when downloading

tj
TJ from which to get data (only works with TJSP for now)

date_start
Lower bound for date

date_end
Upper bound for date

Value

A character vector with the paths to the downloaded files

See also

cjsg_table(), browse_table()